libstdc++: Fix some -Wsystem-headers warnings (PR 95765)
	PR libstdc++/95765
	* include/bits/stl_algobase.h (__size_to_integer(float))
	(__size_to_integer(double), __size_to_integer(long double))
	(__size_to_integer(__float128)): Cast return type explicitly.

(cherry picked from commit 5b6215083bd6a3e10dd142e1c5d4fab011d6f074)

+++ libstdc++-v3/include/bits/stl_algobase.h
@@ -984,14 +984,14 @@ _GLIBCXX_END_NAMESPACE_CONTAINER
 #endif
 
   inline _GLIBCXX_CONSTEXPR long long
-  __size_to_integer(float __n) { return __n; }
+  __size_to_integer(float __n) { return (long long)__n; }
   inline _GLIBCXX_CONSTEXPR long long
-  __size_to_integer(double __n) { return __n; }
+  __size_to_integer(double __n) { return (long long)__n; }
   inline _GLIBCXX_CONSTEXPR long long
-  __size_to_integer(long double __n) { return __n; }
+  __size_to_integer(long double __n) { return (long long)__n; }
 #if !defined(__STRICT_ANSI__) && defined(_GLIBCXX_USE_FLOAT128)
   inline _GLIBCXX_CONSTEXPR long long
-  __size_to_integer(__float128 __n) { return __n; }
+  __size_to_integer(__float128 __n) { return (long long)__n; }
 #endif
 
   template<typename _OutputIterator, typename _Size, typename _Tp>
